requirements.txt
llamaapi
python-dotenv